+++ src/senaite/core/config/fields.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+AT_LABEL_FIELD = "Labels"
+DX_LABEL_FIELD = "Labels"