+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.5.0 (unreleased)
 ------------------
 
+- #2276 Senaite labels
 - #2275 Fix wrong result when both "Result options" and "String" are enabled
 - #2273 Improve performance for sample listing index
 - #2272 Allow to configure the position of additional value columns